Outsourced create_test_sino function to its own file
