+++ tests/test_alg_fmp.py
@@ -4,30 +4,11 @@ import numpy as np
 
 import radontea
 
-
-def create_test_sino(A, N):
-    """
-    Creates test sinogram.
-    """
-    resar = np.zeros((A, N), dtype=float)
-    angles = np.linspace(0, np.pi, A, endpoint=False)
-    x = np.linspace(-N/2, N/2, N, endpoint=True)
-    dev = np.sqrt(N/2)
-    off = N/7
-    for ii in range(A):
-        # Gaussian distribution sinogram
-        x0 = np.cos(angles[ii])*off
-        y = np.exp(-(x-x0)**2/dev**2)
-        resar[ii] = y
-
-    # some normalization, so that max of Radon-inversion
-    # is about 1.
-    resar = 2 * dev * resar / resar.max()
-    return resar, angles
+import sinogram
 
 
 def test_2d_fmp():
-    sino, angles = create_test_sino(A=100, N=101)
+    sino, angles = sinogram.create_test_sino(A=100, N=101)
     r = radontea.fourier_map(sino, angles)
 
     # np.savetxt('outfile.txt', np.array(r).flatten().view(float), fmt="%.8f")
